Extract unused query runner partial
